deleting records

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,26 +1,19 @@
 #!/usr/bin/python3
-"""Module that deletes states containing the letter\
-        "a" from a MySQL database using SQLAlchemy."""
-import sys
+"""Ascript that deletes objects from database"""
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from model_state import State
+import sys
 
 if __name__ == "__main__":
-    # Create the SQLAlchemy engine using the provided MySQL credentials
+    """Creating database engine"""
     engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
                            .format(sys.argv[1], sys.argv[2], sys.argv[3]),
                            pool_pre_ping=True)
-
-    # Create a session factory
     Session = sessionmaker(bind=engine)
-    # Create a session object
-    session = Session()
-    # Retrieve all states from the database
+    session = Session()i
     for state in session.query(State):
-        # Check if the state's name contains the letter "a"
-        if "a" in state.name:
-            # Delete the state from the session
-            session.delete(state)
-    # Commit the session to persist the changes
+        if state is not None:
+            if "a" in state.name:
+                session.delete(state)
     session.commit()
